Revert "[Payment] decouple Payment from Payum and consider decimal factor"
