+++ src/CoreShop/Bundle/PaymentBundle/composer.json
@@ -20,6 +20,7 @@
   "require": {
     "php": "^7.2",
     "coreshop/payment": "^2.0",
+    "coreshop/currency-bundle": "^2.0",
     "coreshop/resource-bundle": "^2.0",
     "coreshop/workflow-bundle": "^2.0",
     "pimcore/pimcore": "^6.6.0"
