--- src/CoreShop/Bundle/PayumBundle/Action/ExecuteSameRequestWithPaymentDetailsAction.php
@@ -1,58 +1,0 @@
-<?php
-/**
- * CoreShop.
- *
- * This source file is subject to the GNU General Public License version 3 (GPLv3)
- * For the full copyright and license information, please view the LICENSE.md and gpl-3.0.txt
- * files that are distributed with this source code.
- *
- * @copyright  Copyright (c) 2015-2019 Dominik Pfaffenbauer (https://www.pfaffenbauer.at)
- * @license    https://www.coreshop.org/license     GNU General Public License version 3 (GPLv3)
- */
-
-namespace CoreShop\Bundle\PayumBundle\Action;
-
-use CoreShop\Component\Core\Model\PaymentInterface;
-use Payum\Core\Action\ActionInterface;
-use Payum\Core\Bridge\Spl\ArrayObject;
-use Payum\Core\Exception\RequestNotSupportedException;
-use Payum\Core\GatewayAwareInterface;
-use Payum\Core\GatewayAwareTrait;
-use Payum\Core\Request\Generic;
-
-final class ExecuteSameRequestWithPaymentDetailsAction implements ActionInterface, GatewayAwareInterface
-{
-    use GatewayAwareTrait;
-
-    /**
-     * {@inheritdoc}
-     *
-     * @param Generic $request
-     */
-    public function execute($request): void
-    {
-        RequestNotSupportedException::assertSupports($this, $request);
-
-        /** @var PaymentInterface $payment */
-        $payment = $request->getModel();
-        $details = ArrayObject::ensureArrayObject($payment->getDetails());
-
-        try {
-            $request->setModel($details);
-
-            $this->gateway->execute($request);
-        } finally {
-            $payment->setDetails((array)$details);
-        }
-    }
-
-    /**
-     * {@inheritdoc}
-     */
-    public function supports($request): bool
-    {
-        return
-            $request instanceof Generic &&
-            $request->getModel() instanceof PaymentInterface;
-    }
-}
