+++ src/CoreShop/Bundle/PayumBundle/Resources/config/services.yml
@@ -26,26 +26,6 @@ services:
         tags:
             - { name: payum.action, all: true, alias: coreshop.confirm_order }
 
-    coreshop.payum_action.authorize_payment: '@CoreShop\Bundle\PayumBundle\Action\AuthorizePaymentAction'
-    CoreShop\Bundle\PayumBundle\Action\AuthorizePaymentAction:
-        arguments:
-            - '%coreshop.currency.decimal_factor%'
-        tags:
-            - { name: payum.action, all: true, alias: coreshop.authorize_payment }
-
-    coreshop.payum_action.capture_payment: '@CoreShop\Bundle\PayumBundle\Action\CapturePaymentAction'
-    CoreShop\Bundle\PayumBundle\Action\CapturePaymentAction:
-        arguments:
-            - '%coreshop.currency.decimal_factor%'
-        tags:
-            - { name: payum.action, all: true, alias: coreshop.capture_payment }
-
-    coreshop.payum_action.execute_same_request_with_payment_details: '@CoreShop\Bundle\PayumBundle\Action\ExecuteSameRequestWithPaymentDetailsAction'
-    CoreShop\Bundle\PayumBundle\Action\ExecuteSameRequestWithPaymentDetailsAction:
-        tags:
-            - { name: payum.action, all: true }
-
-
     # Offline Payment
     coreshop.payum_action.offline.convert_payment: '@CoreShop\Bundle\PayumBundle\Action\Offline\ConvertPaymentAction'
     CoreShop\Bundle\PayumBundle\Action\Offline\ConvertPaymentAction:
