+++ src/CoreShop/Component/Core/Model/PaymentInterface.php
@@ -16,5 +16,4 @@ use CoreShop\Component\Order\Model\OrderPaymentInterface;
 
 interface PaymentInterface extends OrderPaymentInterface
 {
-
 }
