+++ src/CoreShop/Component/Payment/composer.json
@@ -18,7 +18,8 @@
     }
   ],
   "require": {
-    "php": "^7.2"
+    "php": "^7.2",
+    "coreshop/currency": "^2.0"
   },
   "config": {
     "bin-dir": "bin"
